Moving power BI export to separate controller
