+++ app/controllers/reports_controller.rb
@@ -62,27 +62,7 @@ class ReportsController < ApplicationController
     end
   end
 
-  def power_bi
-    authorize :report, :power_bi?
-    @form = Forms::FinanceReport.new
-  end
-
-  def power_bi_export
-    authorize :report, :power_bi?
-    send_file power_bi_data_file
-  end
-
   private
-
-  def power_bi_data_file
-    power_bi = Views::Reports::PowerBiExport.new
-    power_bi.zipfile_path
-  rescue StandardError => e
-    Raven.tags_context(task: "power_bi_export") do
-      Raven.capture_message(e.message)
-    end
-    Rails.logger.debug "Error in power_bi export task: #{e.message}"
-  end
 
   def form
     Forms::FinanceReport.new(report_params)
