+++ app/views/reports/index.html.slim
@@ -32,5 +32,5 @@ dl.govuk-summary-list
       dd.govuk-summary-list__value Extract income claims data by court and date
   - if policy(:report).power_bi?
     .govuk-summary-list__row
-      dt.govuk-summary-list__key =link_to 'Power BI export', reports_power_bi_path
+      dt.govuk-summary-list__key =link_to 'Power BI export', report_power_bi_path
       dd.govuk-summary-list__value Data export for Power BI software
