+++ app/views/reports/power_bi.html.slim
@@ -3,7 +3,7 @@
 
 h1.govuk-heading-xl = "Power BI export"
 
-= form_for @form, url: reports_power_bi_path, method: :put, html: { autocomplete: 'off' } do |f|
+= form_for @form, url: report_power_bi_path, method: :put, html: { autocomplete: 'off' } do |f|
 
  .govuk-form-group.group-level
     = f.submit 'Generate report', class: 'govuk-button'
