+++ config/routes.rb
@@ -14,14 +14,14 @@ Rails.application.routes.draw do
   get 'letter_templates' => 'reports#letters'
   get 'reports/raw_data' => 'reports#raw_data'
   put 'reports/raw_data' => 'reports#raw_data_export'
-  get 'reports/power_bi' => 'reports#power_bi'
-  put 'reports/power_bi' => 'reports#power_bi_export'
 
   namespace :report do
     get 'ccmcc_data' => 'ccmcc_data#show'
     put 'ccmcc_data' => 'ccmcc_data#data_export'
     get 'income_claims_data' => 'income_claims_data#show'
     put 'income_claims_data' => 'income_claims_data#data_export'
+    get 'power_bi' => 'power_bi#show'
+    put 'power_bi' => 'power_bi#data_export'
   end
 
   get '/applications/new' => 'applications/build#create'
